refactor(aruco): cleaning
