+++ userspace/api/src/applications/modules/aruco/aruco_utils/detection.py
@@ -1,23 +1,16 @@
-from dataclasses import dataclass
-
-import numpy as np
+from typing import Any, List
 
 
-@dataclass
 class DetectFlag:
-    target_id: int = -1
-    flag: bool = False
-    corner: np.array=False
+
+    def __init__(self,target_dict:dict={}) -> None:
+        self.target_dict=target_dict
+        self.flag = False
 
     def __bool__(self):
         return self.flag
 
     def set(self, validate=True):
-        """Sets flag. target_id must be set before calling this fuction if validate is True (default)."""
-
-        if validate and self.target_id == -1:
-            raise ValueError("target_id must be set before setting flag.")
-
         self.flag = True
 
     def unset(self):
@@ -25,11 +18,9 @@ class DetectFlag:
 
     def toggle(self):
         self.flag = not self.flag
-    
-    def update(self,new_corner):
-        self.corner=new_corner
 
-    def change_target(self, new_target: int):
-        """Change the target_id"""
-        self.target_id = new_target
-        self.unset()
+    def update(self,key:int, value:Any)->None:
+        self.target_dict.update({key:value})
+
+    def target_exist(self, target_id:int)->bool:
+        return target_id in self.target_dict.keys()
